+++ examples/client-app/app.js
@@ -1,10 +1,7 @@
 // import splitster from "splitster"
 import Cookies from "js-cookie";
 import R from "ramda";
-import {
-  parseCookies,
-  initSplitsterBrowser as splitsterInit
-} from "../../src/index";
+import { parseCookies, init as initSplitster } from "../../src/index";
 
 import config, { BUTTON_COLOR } from "./config";
 
@@ -13,8 +10,8 @@ const user = {
 };
 const init = () => {
   const userId = document.getElementById("userid").value;
-  const splitster = splitsterInit(config, user, userId, {
-    BUTTON_COLOR_1: "red"
+  const splitster = initSplitster(config, user, userId, {
+    BUTTON_COLOR_1: "blue"
   });
   window.splitster = splitster;
   console.log("values:", splitster.getSaveResults());
