--- src/clients/browser/index.js.flow
@@ -1,13 +1,0 @@
-// @flow strict
-import type { Config } from "../../records/config";
-import type { SplitsterResults } from "../index";
-import type { SplitsterBrowser } from "./SplitsterBrowser";
-
-export type init = (
-  config: Config,
-  user: Object,
-  userId: string,
-  override?: SplitsterResults
-) => SplitsterBrowser;
-
-declare export default init;
