--- src/clients/server/SplitsterServer.js
@@ -1,8 +1,0 @@
-import * as R from "ramda";
-
-import SplitsterClient from "../index";
-import { getTestsFromConfig } from "../../records/test";
-
-class SplitsterServer extends SplitsterClient {}
-
-export default SplitsterServer;
