--- src/clients/server/SplitsterServer.js.flow
@@ -1,26 +1,0 @@
-// @flow strict
-import type { Tests } from "../../records/test";
-import type { Config, OptionsConfig } from "../../records/config";
-import type { SplitsterResults, ClientInput, VariantValue } from "../index";
-
-import SplitsterClient from "../index";
-
-declare class SplitsterServerClass extends SplitsterClient {
-  options: OptionsConfig;
-
-  user: string;
-
-  result: SplitsterResults;
-
-  constructor(input: ClientInput, copy: ClientInput): SplitsterClient;
-
-  getSaveResults(): SplitsterResults;
-
-  get(testId: string): VariantValue;
-
-  set(testId: string, variantId: string, cookie?: boolean): Client;
-}
-
-export type SplitsterServer = typeof SplitsterServerClass;
-
-declare export default SplitsterServer;
