--- src/clients/server/index.js.flow
@@ -1,13 +1,0 @@
-// @flow strict
-import type { Config } from "../../records/config";
-import type { SplitsterResults } from "../index";
-import type { SplitsterServer } from "./SplitsterServer";
-
-export type init = (
-  config: Config,
-  user: Object,
-  userId: string,
-  override?: SplitsterResults
-) => SplitsterServer;
-
-declare export default init;
