+++ src/index.js
@@ -1,4 +1,3 @@
-export { default as initSplitsterBrowser } from "./clients/browser";
-export { default as initSplitsterServer } from "./clients/server";
+export { default as init } from "./clients/index";
 
 export { parseCookies } from "./utils/cookies";
